# docs/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Technostrelka API
  version: 1.0.0
  description: |
    Client-server API for Technostrelka Android app (FastAPI + JWT + PostgreSQL).
servers:
  - url: /api/v1
paths:
...
  /users/me:
    get:
      tags: [Users]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeOut'
    patch:
      tags: [Users]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPatchRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeOut'
  /users/me/avatar:
    post:
      tags: [Users]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeOut'
...
  /events/{event_id}/reviews:
    get:
      tags: [Reviews]
      parameters:
        - name: event_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewOut'
  /events/{event_id}/reviews/me:
    put:
      tags: [Reviews]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: event_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpsertRequest'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewOut'
    delete:
      tags: [Reviews]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: event_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: No Content
  /events/{event_id}/reviews/me/photos:
    post:
      tags: [Reviews]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: event_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewOut'
  /review-photos/{photo_id}:
    delete:
      tags: [Reviews]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: photo_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: No Content
...
  /routes:
    get:
      tags: [Routes]
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: goal
          in: query
          schema: { type: string }
        - name: days_range
          in: query
          schema: { type: string }
        - name: people_count
          in: query
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RouteOut'
  /routes/{route_id}:
    get:
      tags: [Routes]
      parameters:
        - name: route_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteDetailOut'
  /routes/recommended/me:
    get:
      tags: [Routes]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteOut'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserMeOut:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        username: { type: string }
        avatar_url: { type: string, nullable: true }
        category_user: { type: string, nullable: true }
    UserPatchRequest:
      type: object
      properties:
        username: { type: string }
        category_user: { type: string }
...
    ReviewOut:
      type: object
      properties:
        id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        user_name: { type: string }
        avatar_url: { type: string, nullable: true }
        rating: { type: integer }
        text: { type: string }
        date: { type: string }
        photo_urls:
          type: array
          items: { type: string }
    ReviewUpsertRequest:
      type: object
      required: [rating, text]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        text: { type: string }
...
    RouteOut:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        image_url: { type: string, nullable: true }
        category: { type: string, nullable: true }
        goal: { type: string, nullable: true }
        days_range: { type: string, nullable: true }
        people_count: { type: string, nullable: true }
        duration: { type: string, nullable: true }
        difficulty: { type: string, nullable: true }
    RouteDetailOut:
      allOf:
        - $ref: '#/components/schemas/RouteOut'
        - type: object
          properties:
            event_ids:
              type: array
              items: { type: string, format: uuid }
...